Simplify simulation code

+++ backend/dates/models.py
@@ -45,8 +45,8 @@ class PossibleDateSet(models.Model):
         Country, related_name="PossibleDateSet", on_delete=models.CASCADE
     )
 
-    n_days_first_restrictions = models.FloatField(null=True)
-    n_days_lockdown = models.FloatField(null=True)
+    n_days_first_restrictions = models.IntegerField(null=True)
+    n_days_lockdown = models.IntegerField(null=True)
     dates_counterfactual_first_restrictions = models.DateField(null=True)
     dates_counterfactual_lockdown = models.DateField(null=True)
